1.4 Update
1st Commit

1.4 Update 2

2nd commit

+++ Banking/Rewards/TileKey.cs
@@ -56,7 +56,7 @@ namespace Banking.Rewards
 					Type = tileOrWallId;
 					break;
 				case TileSubTarget.Wall:
-					Wall = (byte)tileOrWallId;
+					Wall = tileOrWallId;
 					break;
 			}
 		}
